# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a webservice that allows you to create a receipt and calculate a point value of that receipt using a set of rules.",
        "title": "Fetch Receipt Processor API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/receipts": {
            "get": {
                "description": "Get all of the receipts, no limit, no pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reciepts"
                ],
                "summary": "Get All Receipts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.Receipt"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/receipts/process": {
            "post": {
                "description": "Create a receipt and add it to our memory array of receipts. Will not save the id if given one and will always make a new one.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reciepts"
                ],
                "summary": "Process Receipt",
                "parameters": [
                    {
                        "description": "new receipt to create",
                        "name": "receipt",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Receipt"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Returns the ID assigned to the receipt",
                        "schema": {
                            "$ref": "#/definitions/api.CreatedReceiptResponse"
                        }
                    },
                    "400": {
                        "description": "The receipt is invalid",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorMessage"
                        }
                    }
                }
            }
        },
        "/receipts/{id}": {
            "get": {
                "description": "Get the receipt by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reciepts"
                ],
                "summary": "Get A Receipt",
                "parameters": [
                    {
                        "type": "string",
                        "description": "get receipt by id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/models.Receipt"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorMessage"
                        }
                    }
                }
            }
        },
        "/receipts/{id}/points": {
            "get": {
                "description": "Returns the points awarded for the receipt",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reciepts"
                ],
                "summary": "Receipt Points",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The ID of the receipt",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "The number of points awarded",
                        "schema": {
                            "$ref": "#/definitions/api.ReceiptPointsResponse"
                        }
                    },
                    "404": {
                        "description": "No receipt found for that id",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorMessage"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.CreatedReceiptResponse": {
            "description": "Receipt processed response with id",
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "description": "The new receipt id",
                    "type": "string"
                }
            }
        },
        "api.ErrorMessage": {
            "description": "Error Message Information",
            "type": "object",
            "properties": {
                "message": {
                    "description": "The message",
                    "type": "string"
                }
            }
        },
        "api.ReceiptPointsResponse": {
            "description": "Receipt points awarded response with points",
            "type": "object",
            "required": [
                "points"
            ],
            "properties": {
                "points": {
                    "description": "The points awarded for the receipt",
                    "type": "integer"
                }
            }
        },
        "models.Item": {
            "type": "object",
            "required": [
                "price",
                "shortDescription"
            ],
            "properties": {
                "price": {
                    "description": "The total price payed for this item.",
                    "type": "string"
                },
                "shortDescription": {
                    "description": "The Short Product Description for the item.",
                    "type": "string"
                }
            }
        },
        "models.Receipt": {
            "type": "object",
            "required": [
                "items",
                "purchaseDate",
                "purchaseTime",
                "retailer",
                "total"
            ],
            "properties": {
                "id": {
                    "description": "The ID of the receipt",
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Item"
                    }
                },
                "purchaseDate": {
                    "description": "The date of the purchase printed on the receipt.",
                    "type": "string"
                },
                "purchaseTime": {
                    "description": "The time of the purchase printed on the receipt. 24-hour time expected.",
                    "type": "string"
                },
                "retailer": {
                    "description": "The name of the retailer or store the receipt is from.",
                    "type": "string"
                },
                "total": {
                    "description": "The total amount paid on the receipt.",
                    "type": "string"
                }
            }
        }
    }
}